rename /common/SomaticMutation to /common/OpenPedCanSomaticMutations
